server/requirements.txt
asgiref==3.6.0
black==23.1.0
Brotli==1.0.9
click==8.1.3
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
